--story=1002009 --user=林邦泉 定时作业迁移 https://www.tapd.cn/54247054/s/1035938

+++ src/main/java/codedriver/framework/server/core/ServerManager.java
@@ -99,7 +99,7 @@ public class ServerManager implements ApplicationListener<ContextRefreshedEvent>
 		boolean returnVal = false;
 		try {
 			ServerClusterVo serverVo = serverMapper.getServerByServerId(serverId);
-			if (ServerClusterVo.STARTUP.equals(serverVo.getStatus())) {
+			if (serverVo != null && ServerClusterVo.STARTUP.equals(serverVo.getStatus())) {
 				serverVo.setStatus(ServerClusterVo.STOP);
 				serverMapper.updateServerByServerId(serverVo);
 				serverMapper.deleteCounterByServerId(serverVo.getServerId());
